% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check.packages}
\alias{check.packages}
\title{Check if required packages are installed and instll those that are needed}
\usage{
check.packages(pkg)
}
\arguments{
\item{pkg}{required packages as string}
}
\description{
Check if required packages are installed and instll those that are needed
}
\examples{
check.packages(c('ggplot2', 'reshape2'))
}
