% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check.pkg.FIU}
\alias{check.pkg.FIU}
\title{check pkg wrapper
Exception handling to make the installation of ZooScatR and other dependencies easier}
\usage{
check.pkg.FIU()
}
\description{
check pkg wrapper
Exception handling to make the installation of ZooScatR and other dependencies easier
}
